# Improve cache hits for tuple keys in `key_split` and intern results